[refactor](Framework): cast index count in gfx::d3d12::Mesh::draw to UINT for removing compiler warning

+++ common/src/d3d12mesh.cpp
@@ -18,7 +18,7 @@ void Mesh::bind(ID3D12GraphicsCommandList* pCmdList) const {
 
 void Mesh::draw(ID3D12GraphicsCommandList* pCmdList) const {
     DX_THROW_FAILED_VOID( pCmdList->DrawIndexedInstanced(
-        ib().size(), 1, 0, 0, 0
+        static_cast<UINT>( ib().size() ), 1, 0, 0, 0
     ) );
 }
